Implement image/draw.Image as the interface for *ui.Image.
Instead of screwing around with my own experiments, just implement
image/draw.Image. It is significantly slower, but normal people
won't be regenerating large images all the time.

At some point in the future it might make more sense to invent a
fast interface, but for now just make it as simple as possible to
work with.